pyproject.toml
[project]
name = "python-coin-flip-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mitchallen-coin>=0.1.2",
]
